+++ shared-components/statviz/components/filter/constants.ts
@@ -1,0 +1,55 @@
+import { IFilterValue } from "./ValueFilter";
+
+export const genders: IFilterValue[] = [
+  {
+    value: "Boy",
+    label: "Boy",
+    urlId: "boy",
+  },
+  {
+    value: "Girl",
+    label: "Girl",
+    urlId: "girl",
+  },
+  {
+    value: "Men",
+    label: "Men",
+    urlId: "men",
+  },
+  {
+    value: "None",
+    label: "None",
+    urlId: "none",
+  },
+  {
+    value: "Teen Boy",
+    label: "Teen Boy",
+    urlId: "tb",
+  },
+  {
+    value: "Teen Girl",
+    label: "Teen Girl",
+    urlId: "tg",
+  },
+  {
+    value: "UnisexAdult",
+    label: "UnisexAdult",
+    urlId: "ua",
+  },
+  {
+    value: "Unisex Baby",
+    label: "Unisex Baby",
+    urlId: "ub",
+  },
+  {
+    value: "Unisex Kid",
+    label: "Unisex Kid",
+    urlId: "uk",
+  },
+  {
+    value: "Women",
+    label: "Women",
+    urlId: "women",
+  },
+];
+
