+++ shared-components/statviz/components/visualizations/stock/StockDataContainer.tsx
@@ -1,8 +1,3 @@
-import { useQuery } from "@apollo/client";
-import { Box, Spinner } from "@chakra-ui/react";
-import { useParams } from "react-router-dom";
-import StockDataFilter from "./StockDataFilter";
-import ErrorCard, { predefinedErrors } from "../../ErrorCard";
 import { graphql } from "../../../../../graphql/graphql";
 import { TAG_FRAGMENT } from "../../../queries/fragments";
 
@@ -43,21 +38,3 @@ export const STOCK_QUERY = graphql(
   `,
   [TAG_FRAGMENT],
 );
-
-export default function StockDataContainer() {
-  const { baseId } = useParams();
-  const { data, loading, error } = useQuery(STOCK_QUERY, {
-    variables: { baseId: parseInt(baseId!, 10) },
-  });
-
-  if (error) {
-    return <Box>An unexpected error happened {error.message}</Box>;
-  }
-  if (loading) {
-    return <Spinner />;
-  }
-  if (data === undefined) {
-    return <ErrorCard error={predefinedErrors.noData} />;
-  }
-  return <StockDataFilter stockOverview={data.stockOverview!} />;
-}
